adding check for resource lock

+++ python/tests/utils/clean_up_test_resources.py
@@ -13,7 +13,11 @@ def tdr_resource_cleanup() -> None:
     if dataset_info:
         print(f"dataset info found: {dataset_info}")
         dataset_id = dataset_info[0]['id']
-        tdr_client.delete_dataset(dataset_id=dataset_id)
+        if not dataset_info[0].get('resourceLocks'):
+            tdr_client.delete_dataset(dataset_id=dataset_id)
+        else: 
+            lock_id = dataset_info[0]['resourceLocks']['exclusive']
+            tdr_client.unlock_dataset(dataset_id=dataset_id, lock_id=lock_id)
 
 
 if __name__ == "__main__":
